requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
chromadb
pypdf
openai
python-dotenv
streamlit
requests
